modified filenames so / renderes home
